Simplified file names and refactored badge class names
