+++ client-app/src/desktop/tabs/containers/tab/tabs/TabStateExample.js
@@ -5,10 +5,10 @@ import {creates, hoistCmp, HoistModel, managed} from '@xh/hoist/core';
 import {hbox, hspacer} from '@xh/hoist/cmp/layout';
 import {switchInput, textInput} from '@xh/hoist/desktop/cmp/input';
 import {bindable, makeObservable} from '@xh/hoist/mobx';
-import {createContainerModelConfig} from '../TabPanelContainerPanel';
+import {createContainerModelConfig} from './SimpleExample';
 import {find} from 'lodash';
 
-export const tabStateContainerPanel = hoistCmp.factory({
+export const tabStateExample = hoistCmp.factory({
     model: creates(() => new Model()),
 
     render({model}) {
